# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NCH Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: assignedTo
          in: query
          schema:
            type: string
            enum: [mike, jack, jacob]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, done]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task or work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tasks/{taskId}/complete:
    put:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task complete
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}/comments:
    get:
      operationId: listTaskComments
      tags: [tasks]
      summary: List comments on a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskComment"
    post:
      operationId: createTaskComment
      tags: [tasks]
      summary: Add a comment to a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskComment"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        assignedTo:
          type: string
          enum: [mike, jack, jacob]
        assignedBy:
          type: string
          enum: [mike, jack, jacob]
          nullable: true
        propertyAddress:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        priority:
          type: string
          enum: [urgent, normal, low]
        status:
          type: string
          enum: [pending, in_progress, done]
        linkedJobId:
          type: integer
          nullable: true
        linkedJobNumber:
          type: string
          nullable: true
        createdBy:
          type: string
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        isOverdue:
          type: boolean
        commentCount:
          type: integer
      required:
        - id
        - title
        - assignedTo
        - priority
        - status
        - createdBy
        - createdAt
        - isOverdue
        - commentCount
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedTo:
          type: string
          enum: [mike, jack, jacob]
        dueDate:
          type: string
          format: date
        priority:
          type: string
          enum: [urgent, normal, low]
        propertyAddress:
          type: string
        linkedJobId:
          type: integer
      required:
        - title
        - assignedTo
        - priority
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedTo:
          type: string
          enum: [mike, jack, jacob]
        dueDate:
          type: string
          format: date
        priority:
          type: string
          enum: [urgent, normal, low]
        status:
          type: string
          enum: [pending, in_progress, done]
        propertyAddress:
          type: string
        linkedJobId:
          type: integer
    TaskComment:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        authorRole:
          type: string
          enum: [mike, jack, jacob]
        authorName:
          type: string
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - taskId
        - authorRole
        - authorName
        - comment
        - createdAt
    CreateTaskCommentBody:
      type: object
      properties:
        comment:
          type: string
      required:
        - comment
...